feat(backend): added models for the booking app
