+++ backend/booking/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BookingConfig(AppConfig):
+    name = 'booking'
